feat(editor-sidebar): add support for editing article meta on mobile
